fic: wallet modal changes
