+++ package.json
@@ -21,7 +21,7 @@
     "@ethersproject/wallet": "5.6.2",
     "@gnosis.pm/safe-apps-web3-react": "0.6.8",
     "@metamask/jazzicon": "2.0.0",
-    "@pangolindex/components": "7.0.2-internal.47f7d30",
+    "@pangolindex/components": "7.1.3-internal.916bde3",
     "@pangolindex/exchange-contracts": "2.1.7",
     "@pangolindex/governance": "1.2.0",
     "@pangolindex/sdk": "5.0.2-internal.6576c3f",
