+++ src/layout/Header/index.tsx
@@ -5,9 +5,13 @@ import {
   useTranslation,
   TokenInfoModal,
   Tokens,
-  useOnClickOutside
+  useOnClickOutside,
+  useWalletModalToggle,
+  useModalOpen as useModalOpenComponents,
+  ApplicationModal as ApplicationModalComponents,
+  WalletModal
 } from '@pangolindex/components'
-import React, { useState, useRef, useMemo } from 'react'
+import React, { useState, useRef, useMemo, useCallback } from 'react'
 import { usePNGCirculationSupply } from '../../hooks'
 import Web3Status from '../../components/Web3Status'
 import LanguageSelection from '../../components/LanguageSelection'
@@ -37,9 +41,10 @@ import { useChainId } from 'src/hooks'
 import { DISCORD_SUPPORT, LEGACY_PAGE, NETWORK_CURRENCY, NETWORK_LABELS } from 'src/constants'
 import { useMedia } from 'react-use'
 import { MobileHeader } from './MobileHeader'
-import { CHAINS } from '@pangolindex/sdk'
+import { CHAINS, Chain } from '@pangolindex/sdk'
 import { useTotalPngEarnedHook } from 'src/state/stake/multiChainsHooks'
 import { useWeb3React } from '@web3-react/core'
+import { useWallet } from 'src/state/user/hooks'
 
 interface Props {
   activeMobileMenu: boolean
@@ -59,10 +64,16 @@ export default function Header({ activeMobileMenu, handleMobileMenu }: Props) {
 
   const [showPngBalanceModal, setShowPngBalanceModal] = useState(false)
   const [openNetworkSelection, setOpenNetworkSelection] = useState(false)
+  const [selectedChain, setSelectedChain] = useState<Chain | undefined>(undefined)
 
   const node = useRef<HTMLDivElement>()
   const open = useModalOpen(ApplicationModal.FARM)
   const toggle = useToggleModal(ApplicationModal.FARM)
+
+  const walletModalOpen = useModalOpenComponents(ApplicationModalComponents.WALLET)
+  const toggleWalletModal = useWalletModalToggle()
+  const [, setWallet] = useWallet()
+
   useOnClickOutside(node, open ? toggle : undefined)
 
   const [isDark, toggleDarkMode] = useDarkModeManager()
@@ -83,6 +94,24 @@ export default function Header({ activeMobileMenu, handleMobileMenu }: Props) {
 
   const { data: pngCirculationSupply } = usePNGCirculationSupply()
 
+  const handleSelectChain = useCallback(
+    (chain: Chain) => {
+      console.log('chain', chain)
+      setOpenNetworkSelection(false)
+      setSelectedChain(chain)
+      toggleWalletModal()
+    },
+    [setOpenNetworkSelection, setSelectedChain, toggleWalletModal]
+  )
+
+  const onWalletConnect = useCallback(
+    connectorKey => {
+      toggleWalletModal()
+      setWallet(connectorKey)
+    },
+    [setWallet, toggleWalletModal]
+  )
+
   return (
     <HeaderFrame>
       {isMobile ? (
@@ -100,7 +129,11 @@ export default function Header({ activeMobileMenu, handleMobileMenu }: Props) {
               </SupportButton>
             </LegacyButtonWrapper>
             <Hidden upToSmall={true}>
-              <NetworkSelection open={openNetworkSelection} closeModal={closeNetworkSelection} />
+              <NetworkSelection
+                open={openNetworkSelection}
+                closeModal={closeNetworkSelection}
+                onSelectChain={handleSelectChain}
+              />
               {chainId && NETWORK_LABELS[chainId] && (
                 <NetworkCard
                   title={NETWORK_LABELS[chainId]}
@@ -145,6 +178,13 @@ export default function Header({ activeMobileMenu, handleMobileMenu }: Props) {
         circulationSupply={pngCirculationSupply}
         closeModal={closePngBalanceModal}
         token={png}
+      />
+
+      <WalletModal
+        open={walletModalOpen}
+        closeModal={toggleWalletModal}
+        onWalletConnect={onWalletConnect}
+        initialChainId={selectedChain?.chain_id}
       />
     </HeaderFrame>
   )
